Fix trait implementations when Rx/TxToken contains lifetimes
The problem was that the DR/DT types are concrete types with a specific lifetime, so that the where bound required that the associated types of D are always of type DR/DT for all possible lifetimes 'b. This is not the case since the lifetimes of the associated types depend on the trait lifetime. Thus, the device trait was not implemented for some nested types.

(For a minimal code example demonstrating the problem, see [1] [2] [3].)

[1]: https://play.rust-lang.org/?gist=c2e947e3c18a983042e2fb44749ace89&version=stable
[2]: https://play.rust-lang.org/?gist=2da7d2f3d9ffdade1079bc102ed3ba83&version=stable
[3]: https://play.rust-lang.org/?gist=23f5bdd67b75fafe7e4e8cedcc9185ce&version=stable