Fix overflow for send offset and wrong sequence number
During the retransmission, higher sequence numbers than the
last one calculated after (re)sending could be acknowledged.
This led to an overflow in the offset calculation, which
resulted in no further data being sent.
Also, the wrong sequence number would have been used for
following packets.
Now the sequence number to be used for sending is updated
accordingly when the local sequence number is changed.