tcp: implement receiving data on half-closed connection
In the FIN-WAIT-2 state, we have closed the transmit side of our
connection with acknoledgement from the remote side, but we may still
receive data from the remote side. The motivating example is a simple
request/response protocol where the client sends a RST immediately
after its request before/during the server's response.