Add an error buffer to the udp socket
This is a ring buffer storing the failure states of previously received
(and dropped) packet. The buffer will always overwrite the oldest errors
when it is full. An empty buffer, for example Default constructed, will
simply drop all errors. The buffer will keep a count of the number of
ignored errors.

Overall, this makes it possible to indicate more than one failure
condition per call to the library user. This is necessary to not
terminate a `poll` call immediately on the first error. Stored errors
can later be retrieved and inspected outside and independently of the
call that created them.

An instance of the buffer is added to the udp socket but largely unused
at the moment.