Store eth parsing errors in the ethernet interface
Here is why a buffer for errors is needed: The `poll` method of the
ethernet interface may potentially produce more than one signalling
error. But breaking out of control flow leaves the user without a proper
continuation option. A renewed call to `poll` would instead start
iterating sockets and packet flow in the previous order again, giving
unfair advantage to the flows called first. Specifically, the inbound
flow may also be controlled by a malicious party and error during
parsing and reception must not affect other flows.

Instead, we will try to store these errors inside a buffer provided in
the builder so that the programmer can control the number of errors they
want to inspect. Overflowing errors are dropped entirely.

This is part of a patch series introducing that error handling to all
socket types.