Adjust the tests to the udp processing interface
Where tests previously asserted that `process` completed without an
error they will no call process and assert that no error has been
created in the error buffer afterwards. The check does not depend on an
error buffer with non-zero capacity. But when the check has previously
checked for an error case, add an actual error buffer with capacity for
a single element and assert the same condition on the result of `pop`.
Since tests are short enough, the result of these operations can not
wrongly report the error condition of a previous packet processing.