icmpv4 source address is missing
In the latest version of smoltcp there is an assert on line 2420 of
interfaec.rs which requires that ip_repr.src_adr().is_unspecified() be
false.

This contradicts the construction of the icmp v4 packet in the dispatch()
function of icmp.rs, where the src_addr is set to unspecified.

This patch papers over the issue by allowing a calling function
to bind the source ipv4 address.

This is almost certainly _not_ the way it should be done, but
this is more of a PoC to confirm that this is in fact an issue
and this patch fixes it. I'm not quite sure what the most
consistent way is to fix this in the context of the library's APIs.