Add support for sending and receiving DHCP options
This adds the ability to set outgoing DHCP options with an optional buffer,
as well as the option to buffer all incoming DHCP packets and expose them
to the user. Users can then read the relevant information out of the packet
without requiring that support for individual options and fields be compiled
into every usage of smoltcp.

Co-authored-by: Luis-Hebendanz <luis.nixos@gmail.com>