SocketSet: managed slice and socket storage needs seperate lifetimes
Consider a code snippet like below, its compiled with default-features = false
and #![no_std]

```

fn foo<'a>(some_socket: tcp::Socket<'a>) -> Socket<'a> {
    let mut sock_arr: [SocketStorage; 1] = [SocketStorage::EMPTY; 1];
    let onesock: &mut [SocketStorage] = &mut sock_arr;
    let mut sockets = SocketSet::new(onesock);
    let handle = sockets.add(some_socket);
    return sockets.remove(handle);
}
```

The above will fail complaining that &mut sock_arr does not live long enough. This is because
the lifetime of the socket and the slice where the socket is stored are both the same right
now and hence the above cant be done.

```
   |
19 |     let onesock: &mut [SocketStorage] = &mut sock_arr;
   |                                         ------------- `sock_arr` is borrowed here
...
22 |     return sockets.remove(handle);
   |            ^^^^^^^^^^^^^^^^^^^^^^ returns a value referencing data owned by the current function
```

Adding an additional lifetime to make sure that this works fine. And obviously doing this
meant touching all the APIs where SocketSet is used - most of the APIs in interface.rs.
Note that in APIs like poll(), the lifetime 'data is different from lifetime 'a of the
Interface itself (obviously?) because both represent differententities and we dont want
callers of iface.poll(...,socket_set) to have the complaint that socket_set should live
as long as iface