Allow a type implementing `phy::Device` to contain borrowed data
Previously, a limitation of Rust's type system required a device to
implement `phy::Device<'a>` for all lifetimes, including `'static`, to
be able to call methods on `iface::Interface`. This prevented types
implementing `phy::Device<'a>` from containing any borrowed data.

Now that Rust supports GATs (generic associated types), it is possible
to lift this restriction and allow implementations of `phy::Device<'a>`
which contain borrowed data.