wire/udp: make `emit_header` not able to calculate checksum.
Currently it's not working properly, because you can call it when the UDP payload is not written yet to the buffer.
I've changed it to not be able to, and clarified in docs that it's for internal-packet-use only.

It's only used for 6lowpan decompression, and in that case we don't want to calculate and then later check the checksum.