Add implementation of RPL MOP0
This commit is the first commit for adding support for the RPL protocol.
The RPL protocol is a routing protocol used in low-power networks. The
protocol defines 4 modes of operations (MOP0), the routing protocol only
supports upward routes. It is thus not possible to transmit messages
down in to the routing tree.