fix(791): wrong payload length of first IPv4 frag
The payload length of the first IPv4 fragment packet contained the
length of the unfragmented packet. This was because the `repr` was
a clone and not a mutable ref to `ip_repr`. This is now fixed.

We also didn't check that the full IP packet fits in the fragmentation
buffer, which should contain the unfragmented emitted packet.

Fixes #791.

Signed-off-by: Thibaut Vandervelden <thvdveld@vub.be>