refactor: move IPv6 HBH in own module
The Ipv6HopByHopRepr was previously the same as an Ipv6ExtHeader.
However, the RFC says that hop-by-hop options might change, which is not
possible with the current implementation.

I added a representatation for the Hop-by-Hop header, which has an
heapless Vec of parsed options. This way, we can modify the options when
needed.

The function of the Ipv6ExtHeader struct is now purely for parsing the
ext header type and the length. It also returns a pointer to the data it
holds, which must be parsed by the correct extension header
representations.