add support of TCP Segmentation Offloading (TSO)
In case of TSO support, the NIC divides larger data chunks into
TCP segments. To support TSO, the `DeviceCapabilities` is extended
with `TsoCapabilities`. It can describe, if TSO supported for IPv4,
IPv6 or both. In case TSO is supported, smoltcp emits packts with
the maximum size of the remote window to the device driver.