rpl: reset trickle timer when selecting new parent
Resetting the trickle timer when selecting a new parent makes sure that
we send a DIO before we transmit a DAO. This makes sure that all
surrounding nodes have a neighbor cache entry. Otherwise, when
transmitting a DAO before a DIO, a packet might forward it to the root
(in MOP1). The ACK would travel downward using a source routing header.
When there is no entry in the neighbor cache, the intermediate node
sends it back to its parent.