ipv6: add entry to neighbor table
Adding an entry to the neighbor table is very strict and only happens
when we transmit an ARP request or an NDISC request. This now adds
entries in the neighbor table when the destination address is unicast,
in the same network, with a hop limit of 255 or 64 (meaning that the
chance that it is a neighbor is high). We also add it when the
destination address is the ALL_NODES or ALL_RPL_NODES address, as these
packets should not be forwarded.