ipv6: address scope and `is_global_unicast`
The scope of an address is used when selecting the source IPv6 address
based on the destination address. As the scope is then also used for
unicast address, I made the scope only public in the crate instead of
outside the crate. Not sure if this should be public or not.

This commit also adds the `is_global_unicast` query function for IPv6
addresses.