iface: remove always-Ok Results in consume().
They're useless, Ok() was always returned. Additionally, it's misleading
to return a Result from the `consume` closure since it implies TX will
be canceled on Err which is not the case (the driver implementation can't inspect the
result, it just sees a generic `R`, all it can do is pass it through).